+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
 
 ## `v0.1.0`
+
 - duckdb `v1.2.0`
+- linting setup w/ clippy!
+- Added `conn_for_each` and `conn_for_each_blocking` to pool - and tests
+- Made pool auto add read-only config to connections if path is provided in builder otherwise memory...